llm-reasoning-evaluation/requirements.txt
openai==1.23.6
python-dotenv
tqdm
httpx==0.27.2
pandas
torch